1. Bugfix: avoid buffer overflow when the device id is generated by mac.

+++ agent/agent_config.h
@@ -1,11 +1,12 @@
 #ifndef _AGENT_CONFIG_
 #define _AGENT_CONFIG_
 
+#include <stdint.h>
 #include "rmt_agent.h"
 
 typedef struct _agent_config {
     char net_interface[40];          // empty for default interface chosen by RMT agent
-    int device_id;                   // 0 for default device ID generated from MAC
+    uint64_t device_id;              // 0 for default device ID generated from MAC
     int domain_id;                   // 0 for default domain ID 0
     unsigned long datainfo_val_size; // 0 for default size 256
 } agent_config;
